backend-python/requirements.txt
Flask
Flask-CORS
face_recognition
Pillow
numpy